+++ postgres/spec/hyperion/postgres_spec.rb
@@ -44,8 +44,8 @@ describe Hyperion::Postgres do
     context 'Sql Injection' do
       it 'escapes strings to be inserted' do
         evil_name = "my evil name' --"
-        record = Hyperion::API.save(:kind => 'testing', :name => evil_name)
-        found_record = Hyperion::API.find_by_key(record[:key])
+        record = Hyperion.save(:kind => 'testing', :name => evil_name)
+        found_record = Hyperion.find_by_key(record[:key])
         found_record[:name].should == evil_name
       end
 
@@ -53,7 +53,7 @@ describe Hyperion::Postgres do
         evil_name = 'my evil name" --'
         error_message = ""
         begin
-          Hyperion::API.save(:kind => 'my evil name" --', :name => evil_name)
+          Hyperion.save(:kind => 'my evil name" --', :name => evil_name)
         rescue Exception => e
           error_message = e.message
         end
@@ -64,7 +64,7 @@ describe Hyperion::Postgres do
         evil_name = 'my evil name" --'
         error_message = ""
         begin
-          Hyperion::API.save(:kind => 'testing', evil_name => 'value')
+          Hyperion.save(:kind => 'testing', evil_name => 'value')
         rescue Exception => e
           error_message = e.message
         end