+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 * refactors usage of thread local bindings
 * adds entities with fields, timestamps, and defaults
+* collapses Hyperion::API into the root Hyperion module
 
 # 0.0.1.alpha5